+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 1. Added composable to search weather
 2. Add user input to searh weather with latitude and longitude
 3. Add location entity for latitude and longitude
+4. Display weather data as grid
+    1. Add tailwind Css
+    2. Add table layout
+    3. Show More in case more then 1 day of data